Vendor unalix, update ruleset, and change license to LGPLv3
Unalix has not been updated in 3 years, and some URLs such as
https://www.instagram.com/reel/DM8Rv14gvbs/?igsh=MXBldzYweWdkMjg4eA==
are cleaned incorrectly. We vendor unalix and update its ruleset file
with the latest from https://rules2.clearurls.xyz/data.minify.json.

In addition, unalix uses the LGPLv3 license. I do not know if that
requires the complete code to be LGPLv3 or only the unalix/ClearURLs
code. To be safe I am updating my license to LGPLv3.